backend/requirements.txt
fastapi
uvicorn
motor
python-multipart
python-dotenv
emergentintegrations
pydantic
pydantic-settings
